Factor out block processor
